+++ SpeechResponder/ConfigurationWindow.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using EddiConfigService;
 using EddiCore;
+using EddiCore.Hotkeys;
 using EddiEvents;
 using EddiSpeechResponder.ScriptResolverService;
 using EddiSpeechService;
